+++ Content.Shared/_RF/Skills/SharedSkillsSystem.cs
@@ -1,6 +1,9 @@
 using System.Linq;
 using Content.Shared._RF.Skills.Components;
+using Content.Shared.DoAfter;
+using Content.Shared.EntityEffects;
 using Robust.Shared.Prototypes;
+using Robust.Shared.Random;
 
 namespace Content.Shared._RF.Skills;
 
@@ -10,6 +13,7 @@ namespace Content.Shared._RF.Skills;
 public abstract class SharedSkillsSystem : EntitySystem
 {
     [Dependency] protected readonly IPrototypeManager Proto = default!;
+    [Dependency] private readonly IRobustRandom _random = default!;
 
     public int GetLevel(Entity<SkillsComponent?> ent, ProtoId<SkillPrototype> skill)
     {
@@ -57,6 +61,57 @@ public abstract class SharedSkillsSystem : EntitySystem
         return level <= 0 ? 0 : GetLevelMaxPoints(skill, level - 1);
     }
 
+    /// <summary>
+    /// Adds experience points for the skill of the given entity
+    /// </summary>
+    public void AddExperience(Entity<SkillsComponent?> ent, ProtoId<SkillPrototype> skill, int amount)
+    {
+        if (!Resolve(ent, ref ent.Comp, false)
+            || !Proto.TryIndex(skill, out var proto)
+            || ent.Comp.Skills.FirstOrDefault(x => x.Id == skill.Id) is not { } data)
+            return;
+
+        var oldLevel = data.CurrentLevel;
+
+        data.CurrentExp += (int) (amount * data.ExpFactor * ent.Comp.ExpFactor);
+
+        if (data.CurrentExp > data.LevelUpExp && data.CurrentLevel == proto.MaxLevel)
+            data.CurrentExp = data.LevelUpExp;
+
+        if (data.CurrentExp >= data.MinLevelExp && data.CurrentExp <= data.LevelUpExp)
+        {
+            Dirty(ent);
+            return;
+        }
+
+        data.CurrentLevel = GetLevel(data.Id, data.CurrentExp);
+        data.LevelUpExp = GetLevelMaxPoints(data.Id, data.CurrentLevel);
+        data.MinLevelExp = GetLevelMinPoints(data.Id, data.CurrentLevel);
+        Dirty(ent);
+
+        var args = new EntityEffectBaseArgs(ent, EntityManager);
+
+        if (proto.LevelUpEffects.TryGetValue(0, out var zeroEffects))
+        {
+            foreach (var effect in zeroEffects)
+            {
+                if (effect.ShouldApply(args, _random))
+                    effect.Effect(args);
+            }
+        }
+
+        if (data.CurrentLevel != 0 && proto.LevelUpEffects.TryGetValue(data.CurrentLevel, out var effects))
+        {
+            foreach (var effect in effects)
+            {
+                if (effect.ShouldApply(args, _random))
+                    effect.Effect(args);
+            }
+        }
+
+        RaiseLocalEvent(ent, new SkillLevelChanged(data.CurrentLevel, oldLevel));
+    }
+
     #region Interactions
 
     /// <summary>
@@ -78,8 +133,8 @@ public abstract class SharedSkillsSystem : EntitySystem
         var result = delta switch
         {
             0 => value,
-            > 0 => value * delta * ent.Comp.ResultIncreaseFactor,
-            < 0 => value * delta * ent.Comp.ResultDecreaseFactor,
+            > 0 => value + delta * ent.Comp.ResultFactor,
+            < 0 => value - delta * ent.Comp.ResultFactor,
         };
 
         return Math.Clamp(result, ent.Comp.MinResult, ent.Comp.MaxResult);
@@ -92,12 +147,115 @@ public abstract class SharedSkillsSystem : EntitySystem
 
         var delta = GetLevel(user, ent.Comp.Skill) - ent.Comp.TargetLevel;
 
-        return Math.Clamp(delay - delay * delta * ent.Comp.DoAfterFactor,
+        return Math.Clamp(delay - delta * ent.Comp.DoAfterFactor,
             ent.Comp.MinDoAfterTime,
             ent.Comp.MaxDoAfterTime);
     }
 
+    public SkillCheckResult DoInteractionCheck(Entity<SkillInteractionComponent?> ent, DoAfterEvent args)
+    {
+        return DoInteractionCheck(ent, args.User, args.Target);
+    }
+
+    /// <summary>
+    /// Performs skill checks for interaction and, according to the result of the check,
+    /// gives out experience and triggers interaction effects
+    /// </summary>
+    /// <param name="ent">Entity on which the interaction is performed</param>
+    /// <param name="user">The user who performs the interaction</param>
+    /// <param name="target">Target entity of the interaction</param>
+    /// <returns>Skills check result for interaction</returns>
+    public SkillCheckResult DoInteractionCheck(
+        Entity<SkillInteractionComponent?> ent,
+        Entity<SkillsComponent?> user,
+        EntityUid? target)
+    {
+        if (!Resolve(ent, ref ent.Comp, false) || !Resolve(user, ref user.Comp, false))
+            return SkillCheckResult.Success;
+
+        var interact = ent.Comp;
+        var delta = GetLevel(user, interact.Skill) - interact.TargetLevel;
+        var successChance = interact.SuccessFactor * delta + 0.2f;
+        var failChance = 0.4f - successChance;
+
+        var fail = _random.Prob(failChance);
+        var success = _random.Prob(successChance);
+
+        if (success)
+        {
+            AddExperience(user, ent.Comp.Skill, (int)(interact.Experience * interact.ExpSuccessFactor));
+
+            foreach (var effect in interact.SuccessEffects)
+            {
+                var args = new EntityEffectBaseArgs(ent, EntityManager);
+
+                if (effect.ShouldApply(args, _random))
+                    effect.Effect(args);
+            }
+
+            if (target == null)
+                return SkillCheckResult.Fail;
+
+            foreach (var effect in interact.SuccessTargetEffects)
+            {
+                var args = new EntityEffectBaseArgs(target.Value, EntityManager);
+
+                if (effect.ShouldApply(args, _random))
+                    effect.Effect(args);
+            }
+
+            return SkillCheckResult.AdditionalSuccess;
+        }
+
+        if (fail)
+        {
+            AddExperience(user, ent.Comp.Skill, (int)(interact.Experience * interact.ExpFailFactor));
+
+            foreach (var effect in interact.FailEffects)
+            {
+                var args = new EntityEffectBaseArgs(ent, EntityManager);
+
+                if (effect.ShouldApply(args, _random))
+                    effect.Effect(args);
+            }
+
+            if (target == null)
+                return SkillCheckResult.Fail;
+
+            foreach (var effect in interact.FailTargetEffects)
+            {
+                var args = new EntityEffectBaseArgs(target.Value, EntityManager);
+
+                if (effect.ShouldApply(args, _random))
+                    effect.Effect(args);
+            }
+
+            return SkillCheckResult.Fail;
+        }
+
+        AddExperience(user, ent.Comp.Skill, interact.Experience);
+        return SkillCheckResult.Success;
+    }
+
     #endregion
+}
+
+public enum SkillCheckResult : byte
+{
+    /// <summary>
+    /// Interaction successfully completed with additional effects invoked
+    /// </summary>
+    AdditionalSuccess,
+
+    /// <summary>
+    /// Interaction successfully completed
+    /// </summary>
+    Success,
+
+    /// <summary>
+    /// The interaction failed with the invocation of additional effects
+    /// </summary>
+    Fail,
 }
 
 /// <summary>
