+++ Resources/Locale/en-US/_RF/skills/skills.ftl
@@ -1,0 +1,2 @@
+skill-healing-name = Healing
+skill-construction-name = Construction
